move searchBar related logic into searchBar.js
